Edited (first pass)
As is, this article needs to be duplicated for different distributions
(4 articles needed). However, we need to resolve what the subject
matter is. The title is “Configuring and Maintaining Ghost” but the
content mostly covers “Installing and Maintaining Ghost.” We have other
“installing and Maintaining Ghost” articles (e.g., 00030). Maybe check
with GD about what is meant by “maintaining.”